Extract hapi-fhir-repositories module
