+++ hapi-fhir-base/src/main/java/ca/uhn/fhir/repository/Repositories.java
@@ -1,9 +1,7 @@
 package ca.uhn.fhir.repository;
 
 import ca.uhn.fhir.context.FhirContext;
-import ca.uhn.fhir.repository.impl.GenericClientRepository;
-import ca.uhn.fhir.repository.loader.UrlRepositoryFactory;
-import ca.uhn.fhir.rest.client.api.IGenericClient;
+import ca.uhn.fhir.repository.impl.UrlRepositoryFactory;
 import jakarta.annotation.Nonnull;
 import jakarta.annotation.Nullable;
 
@@ -36,7 +34,4 @@ public class Repositories {
 		return UrlRepositoryFactory.buildRepository(theBaseUrl, theFhirContext);
 	}
 
-	public static IRepository restClientRepository(IGenericClient theGenericClient) {
-		return new GenericClientRepository(theGenericClient);
-	}
 }
