+++ hapi-fhir-repositories/src/main/java/ca/uhn/fhir/repository/impl/memory/InMemoryFhirRepositoryLoader.java
@@ -4,7 +4,7 @@ import ca.uhn.fhir.context.FhirContext;
 import ca.uhn.fhir.i18n.Msg;
 import ca.uhn.fhir.repository.IRepository;
 import ca.uhn.fhir.repository.loader.BaseSchemeBasedFhirRepositoryLoader;
-import ca.uhn.fhir.repository.loader.IRepositoryLoader;
+import ca.uhn.fhir.repository.IRepositoryLoader;
 import jakarta.annotation.Nonnull;
 import org.apache.commons.collections4.map.ReferenceMap;
 
@@ -13,7 +13,7 @@ import org.apache.commons.collections4.map.ReferenceMap;
  */
 public class InMemoryFhirRepositoryLoader extends BaseSchemeBasedFhirRepositoryLoader implements IRepositoryLoader {
 
-	// This Loader is registered under META-INF/services/ca.uhn.fhir.repository.loader.IRepositoryLoader
+	// This Loader is registered under META-INF/services/ca.uhn.fhir.repository.IRepositoryLoader
 
 	public static final String URL_SUB_SCHEME = "memory";
 	static final ReferenceMap<String, InMemoryFhirRepository> ourRepositories = new ReferenceMap<>();
