+++ hapi-fhir-repositories/src/main/java/ca/uhn/fhir/repository/impl/memory/ResourceStorage.java
@@ -57,7 +57,6 @@ public class ResourceStorage {
 		return lookupResource(theId);
 	}
 
-	@VisibleForTesting
 	public @Nonnull Collection<IBaseResource> getAllOfType(String theResourceType) {
 		return getResourceMapForType(theResourceType).values();
 	}
