+++ hapi-fhir-repositories/src/main/resources/META-INF/services/ca.uhn.fhir.repository.IRepositoryLoader
